+++ src/packet/smi.rs
@@ -71,6 +71,11 @@ impl<P> Packet<P> {
     pub fn payload(&self) -> &P {
         &self.payload
     }
+
+    /// Get a mutable reference to the packet's payload
+    pub fn payload_mut(&mut self) -> &mut P {
+        &mut self.payload
+    }
 }
 
 impl<U: unit::Unit> Packet<Decoder<'_, U>> {
@@ -105,6 +110,14 @@ where
         } else {
             Ok(res)
         }
+    }
+}
+
+impl<U: unit::Unit> TryFrom<Packet<Decoder<'_, U>>> for payload::Payload<U::IOptions, U::DOptions> {
+    type Error = Error;
+
+    fn try_from(packet: Packet<Decoder<'_, U>>) -> Result<Self, Self::Error> {
+        packet.decode_payload()
     }
 }